merge detection and legal boundaries
blocks recognize merging event, we can start on the merge and validation methods with this new code addition. The blocks are also reset to their starting position if they go outside of the permitted boundaries. Also, Block class has been created with isLeft Boolean variable so it has the attribute of letting us know what side of the equation is on. there is a setIsLeft() method to change the value to true or false and there is also a getIsLeft() method to retrieve the variable. Feel free to add any additional methods that will help with the block attributes. note: merge and validation methods can now be included on the jFrame code.