Quantity Work Started working on quantites, and they are stll pretty rough. Quantities inherit from an abstract base quantity that specifies what operations they need to implement as well as exposes them to outside classes. It also has two vectors, one for holding an unspecified number of blocks, and the other for holding an unspecified number of quantities. Right now add and remove are public, but I think only other quantities should be able to access them. I don't want the equation to be able to add/remove a quantity all willy nilly because it might violate some rule of algebra.
The quantity subclasses all enforce there own ruleset. The Equal
Quantity hold 2 and only 2 quantites. When a block is promoted into the
quantity, it is instead removed from the source, operated on depending
on the source quantity's class, and then added to the other quantity.
Merging and Demoting are not allowed for the Equal Quantity.

The Additon quantity only implements merge currently, where it adds two
blocks it contains if able. It will later implement promote and demote,
but that will take a bit of work since it needs to apply different rules
depending on the other quantity involved.

The basic gist of quantites is that they can manipulate contained
blocks, demote a contained block into a sub-quantity, and promote a
block a sub-quantity contains into itself. Blocks can only be moved one
quantity at a time. The equation class will hold onto the root quantity,
giving it access to all blocks and quantities.