v7 merge() completed and minor changes
merge function has been completed. I had to take away some locks if not all because it was getting too hard to control the flow of the threads. The program still runs in multiple threads aside from the EDT. One limitation is no coefficient with variables, no fraction with variables unless an integer is input as the denominator or numerator. I just couldn't get to that point and I would have to use more than 4 blocks. In that case it would be ideal to have all the attributes in the block class and having them created or deleted. that way we wouldn't mind how many blocks we would need. There is a glitch where the block or blocks move to the farthest position when we merge them. This only happens when we use the swap class. I can't get rid of it since it is a glitch that is coming from the EDT, in that case, if we want to get rid of it, we would have to create the blocks and deleted them so this problem doesn't happen. Moving forward, I will try to give it a simple fix if I have the time. I still haven't study for the final since the debugging took most of my time. I'll see you all tomorrow at 3pm.